Refactor ideascale cli related objects to bin cli
